+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+
+### Added
+
+- Add organization Asset connector fields
+
 ## 1.21.1 - 2025-10-28
 
 ### Added