+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Move to `uv` to build project
+
 ## 1.22.0 - 2025-12-15
 
 ### Added
@@ -83,8 +87,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Add users fields to enrichment 
-- Fix asset connector post url 
+- Add users fields to enrichment
+- Fix asset connector post url
 - Update OCSF fields for software, user and vulnerability
 - Fix some typos
 - Add some extra tests for updating connector fields