+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Fix some OCSF schema fields 
+
 ## 1.22.0 - 2025-12-15
 
 ### Added