+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 1.22.1 - 2026-01-15
+
 ### Fixed
 
 - Fix some OCSF schema fields